requirements.txt
langchain
langchain-ollama
langgraph
beautifulsoup4
requests
psutil
GPUtil
duckduckgo-search
fpdf2
